Refactor Alerts Detail Page out into separate file.
TODO: pull silence UI pieces out into a separate utils file
