+++ src/components/alerting.tsx
@@ -1,16 +1,11 @@
 import {
-  ActionServiceProvider,
   Alert,
   AlertSeverity,
   AlertStates,
   consoleFetchJSON,
-  GreenCheckCircleIcon,
   ListPageFilter,
   PrometheusAlert,
   PrometheusEndpoint,
-  PrometheusLabels,
-  ResourceLink,
-  ResourceStatus,
   RowFilter,
   RowProps,
   Rule,
@@ -30,26 +25,17 @@ import {
   Checkbox,
   CodeBlock,
   CodeBlockCode,
-  Dropdown,
   DropdownItem,
-  DropdownPosition,
   DropdownToggle,
   Flex,
   FlexItem,
-  KebabToggle,
-  Label,
-  Modal,
-  ModalVariant,
-  Popover,
   Toolbar,
   ToolbarContent,
   ToolbarGroup,
   ToolbarItem,
 } from '@patternfly/react-core';
-import { BanIcon, HourglassHalfIcon } from '@patternfly/react-icons';
 import { sortable } from '@patternfly/react-table';
 import classNames from 'classnames';
-import { TFunction } from 'i18next';
 import * as _ from 'lodash-es';
 import * as React from 'react';
 import { Helmet } from 'react-helmet';
@@ -69,18 +55,8 @@ import {
   isAlertingRulesSource,
 } from './console/extensions/alerts';
 import { getPrometheusURL } from './console/graphs/helpers';
-import {
-  ContainerModel,
-  DaemonSetModel,
-  DeploymentModel,
-  JobModel,
-  NamespaceModel,
-  NodeModel,
-  PodModel,
-  StatefulSetModel,
-} from './console/models';
 import { SectionHeading } from './console/utils/headings';
-import { ExternalLink, LinkifyExternal } from './console/utils/link';
+import { ExternalLink } from './console/utils/link';
 import { getAllQueryArguments } from './console/utils/router';
 import { LoadingInline, StatusBox } from './console/utils/status-box';
 
@@ -96,23 +72,19 @@ import { useBoolean } from './hooks/useBoolean';
 import KebabDropdown from './kebab-dropdown';
 import { Labels } from './labels';
 import { QueryBrowserPage, ToggleGraph } from './metrics';
-import { FormatSeriesTitle, QueryBrowser } from './query-browser';
 import { CreateSilence, EditSilence } from './silence-form';
 import { TargetsUI } from './targets';
 import { Alerts, AlertSource, RootState, Silences } from './types';
 import {
   alertDescription,
   alertingRuleStateOrder,
-  AlertResource,
   alertSeverityOrder,
-  alertState,
   alertURL,
   fuzzyCaseInsensitive,
   getAlertsAndRules,
   labelsToParams,
   refreshSilences,
   RuleResource,
-  silenceMatcherEqualitySymbol,
   SilenceResource,
   silenceState,
 } from './utils';
@@ -121,101 +93,36 @@ import './_monitoring.scss';
 import { usePerspective } from './hooks/usePerspective';
 import {
   alertingRuleSource,
-  alertSource,
   AlertState,
-  AlertStateDescription,
   AlertStateIcon,
-  devRuleURL,
   getAdditionalSources,
   getAlertStateKey,
-  isActionWithCallback,
-  isActionWithHref,
+  getSourceKey,
+  Graph,
   MonitoringResourceIcon,
+  OnToggle,
+  PopoverField,
+  queryBrowserURL,
   ruleURL,
+  SelectedSilencesContext,
   Severity,
-  SeverityIcon,
+  SeverityBadge,
+  SeverityCounts,
+  SeverityHelp,
   silenceAlertURL,
+  SilenceDropdown,
+  SilenceMatchersList,
   SilencesNotLoadedWarning,
+  SilenceState,
+  SilenceTableRow,
+  SourceHelp,
+  tableSilenceClasses,
 } from './alerting/AlertUtils';
 import AlertsPage from './alerting/AlertsPage';
-
-const SelectedSilencesContext = React.createContext({
-  selectedSilences: new Set(),
-  setSelectedSilences: undefined,
-});
+import AlertsDetailsPage from './alerting/AlertsDetailPage';
 
 const pollers = {};
 const pollerTimeouts = {};
-
-const SilenceState = ({ silence }) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  const state = silenceState(silence);
-  const icon = {
-    [SilenceStates.Active]: <GreenCheckCircleIcon />,
-    [SilenceStates.Pending]: <HourglassHalfIcon className="monitoring-state-icon--pending" />,
-    [SilenceStates.Expired]: <BanIcon className="text-muted" data-test-id="ban-icon" />,
-  }[state];
-
-  const getStateKey = (stateData) => {
-    switch (stateData) {
-      case SilenceStates.Active:
-        return t('Active');
-      case SilenceStates.Pending:
-        return t('Pending');
-      default:
-        return t('Expired');
-    }
-  };
-
-  return icon ? (
-    <>
-      {icon} {getStateKey(state)}
-    </>
-  ) : null;
-};
-
-const StateTimestamp = ({ text, timestamp }) => (
-  <div className="text-muted monitoring-timestamp">
-    {text}&nbsp;
-    <Timestamp timestamp={timestamp} />
-  </div>
-);
-
-const SeverityBadge: React.FC<{ severity: string }> = React.memo(({ severity }) =>
-  _.isNil(severity) || severity === 'none' ? null : (
-    <ResourceStatus>
-      <Severity severity={severity} />
-    </ResourceStatus>
-  ),
-);
-
-const SeverityCounts: React.FC<{ alerts: Alert[] }> = ({ alerts }) => {
-  if (_.isEmpty(alerts)) {
-    return <>-</>;
-  }
-
-  const counts = _.countBy(alerts, (a) => {
-    const { severity } = a.labels;
-    return severity === AlertSeverity.Critical || severity === AlertSeverity.Warning
-      ? severity
-      : AlertSeverity.Info;
-  });
-
-  const severities = [AlertSeverity.Critical, AlertSeverity.Warning, AlertSeverity.Info].filter(
-    (s) => counts[s] > 0,
-  );
-
-  return (
-    <>
-      {severities.map((s) => (
-        <span className="monitoring-icon-wrap" key={s}>
-          <SeverityIcon severity={s} /> {counts[s]}
-        </span>
-      ))}
-    </>
-  );
-};
 
 const StateCounts: React.FC<{ alerts: PrometheusAlert[] }> = ({ alerts }) => {
   const { t } = useTranslation('plugin__monitoring-plugin');
@@ -236,648 +143,9 @@ const StateCounts: React.FC<{ alerts: PrometheusAlert[] }> = ({ alerts }) => {
   );
 };
 
-const PopoverField: React.FC<{ bodyContent: React.ReactNode; label: string }> = ({
-  bodyContent,
-  label,
-}) => (
-  <Popover headerContent={label} bodyContent={bodyContent}>
-    <Button variant="plain" className="details-item__popover-button">
-      {label}
-    </Button>
-  </Popover>
-);
-
-const AlertStateHelp: React.FC = () => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  return (
-    <dl className="co-inline">
-      <dt>
-        <AlertStateIcon state={AlertStates.Pending} /> <strong>{t('Pending: ')}</strong>
-      </dt>
-      <dd>
-        {t(
-          'The alert is active but is waiting for the duration that is specified in the alerting rule before it fires.',
-        )}
-      </dd>
-      <dt>
-        <AlertStateIcon state={AlertStates.Firing} /> <strong>{t('Firing: ')}</strong>
-      </dt>
-      <dd>
-        {t(
-          'The alert is firing because the alert condition is true and the optional `for` duration has passed. The alert will continue to fire as long as the condition remains true.',
-        )}
-      </dd>
-      <dt>
-        <AlertStateIcon state={AlertStates.Silenced} /> <strong>{t('Silenced: ')}</strong>
-      </dt>
-      <dt></dt>
-      <dd>
-        {t(
-          'The alert is now silenced for a defined time period. Silences temporarily mute alerts based on a set of label selectors that you define. Notifications will not be sent for alerts that match all the listed values or regular expressions.',
-        )}
-      </dd>
-    </dl>
-  );
-};
-
-const SeverityHelp: React.FC = () => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  return (
-    <dl className="co-inline">
-      <dt>
-        <SeverityIcon severity={AlertSeverity.Critical} /> <strong>{t('Critical: ')}</strong>
-      </dt>
-      <dd>
-        {t(
-          'The condition that triggered the alert could have a critical impact. The alert requires immediate attention when fired and is typically paged to an individual or to a critical response team.',
-        )}
-      </dd>
-      <dt>
-        <SeverityIcon severity={AlertSeverity.Warning} /> <strong>{t('Warning: ')}</strong>
-      </dt>
-      <dd>
-        {t(
-          'The alert provides a warning notification about something that might require attention in order to prevent a problem from occurring. Warnings are typically routed to a ticketing system for non-immediate review.',
-        )}
-      </dd>
-      <dt>
-        <SeverityIcon severity={AlertSeverity.Info} /> <strong>{t('Info: ')}</strong>
-      </dt>
-      <dd>{t('The alert is provided for informational purposes only.')}</dd>
-      <dt>
-        <SeverityIcon severity={AlertSeverity.None} /> <strong>{t('None: ')}</strong>
-      </dt>
-      <dd>{t('The alert has no defined severity.')}</dd>
-      <dd>{t('You can also create custom severity definitions for user workload alerts.')}</dd>
-    </dl>
-  );
-};
-
-const SourceHelp: React.FC = () => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  return (
-    <dl className="co-inline">
-      <dt>
-        <strong>{t('Platform: ')}</strong>
-      </dt>
-      <dd>
-        {t(
-          'Platform-level alerts relate only to OpenShift namespaces. OpenShift namespaces provide core OpenShift functionality.',
-        )}
-      </dd>
-      <dt>
-        <strong>{t('User: ')}</strong>
-      </dt>
-      <dd>
-        {t(
-          'User workload alerts relate to user-defined namespaces. These alerts are user-created and are customizable. User workload monitoring can be enabled post-installation to provide observability into your own services.',
-        )}
-      </dd>
-    </dl>
-  );
-};
-
-const queryBrowserURL = (query: string, namespace: string) =>
-  namespace
-    ? `/dev-monitoring/ns/${namespace}/metrics?query0=${encodeURIComponent(query)}`
-    : `/monitoring/query-browser?query0=${encodeURIComponent(query)}`;
-
-const Graph: React.FC<GraphProps> = ({
-  filterLabels = undefined,
-  formatSeriesTitle,
-  namespace,
-  query,
-  ruleDuration,
-}) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  // 3 times the rule's duration, but not less than 30 minutes
-  const timespan = Math.max(3 * ruleDuration, 30 * 60) * 1000;
-
-  const GraphLink = () =>
-    query ? (
-      <Link aria-label={t('View in Metrics')} to={queryBrowserURL(query, namespace)}>
-        {t('View in Metrics')}
-      </Link>
-    ) : null;
-
-  return (
-    <QueryBrowser
-      namespace={namespace}
-      defaultTimespan={timespan}
-      filterLabels={filterLabels}
-      formatSeriesTitle={formatSeriesTitle}
-      GraphLink={GraphLink}
-      pollInterval={Math.round(timespan / 120)}
-      queries={[query]}
-    />
-  );
-};
-
-const tableSilenceClasses = [
-  'pf-c-table__action', // Checkbox
-  'pf-u-w-50 pf-u-w-33-on-sm', // Name
-  'pf-m-hidden pf-m-visible-on-sm', // Firing alerts
-  '', // State
-  'pf-m-hidden pf-m-visible-on-sm', // Creator
-  'dropdown-kebab-pf pf-c-table__action',
-];
-
-const SilenceMatchersList = ({ silence }) => (
-  <div className={`co-text-${SilenceResource.kind.toLowerCase()}`}>
-    {_.map(silence.matchers, ({ name, isEqual, isRegex, value }, i) => (
-      <Label className="co-label" key={i}>
-        <span className="co-label__key">{name}</span>
-        <span className="co-label__eq">{silenceMatcherEqualitySymbol(isEqual, isRegex)}</span>
-        <span className="co-label__value">{value}</span>
-      </Label>
-    ))}
-  </div>
-);
-
-type SilenceTableRowProps = {
-  obj: Silence;
-  showCheckbox?: boolean;
-};
-
-const SilenceTableRow: React.FC<SilenceTableRowProps> = ({ obj, showCheckbox }) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  const { createdBy, endsAt, firingAlerts, id, name, startsAt } = obj;
-  const state = silenceState(obj);
-
-  const { selectedSilences, setSelectedSilences } = React.useContext(SelectedSilencesContext);
-
-  const onCheckboxChange = React.useCallback(
-    (isChecked: boolean) => {
-      setSelectedSilences((oldSet) => {
-        const newSet = new Set(oldSet);
-        if (isChecked) {
-          newSet.add(id);
-        } else {
-          newSet.delete(id);
-        }
-        return newSet;
-      });
-    },
-    [id, setSelectedSilences],
-  );
-
-  return (
-    <>
-      {showCheckbox && (
-        <td className={tableSilenceClasses[0]}>
-          <Checkbox
-            id={id}
-            isChecked={selectedSilences.has(id)}
-            isDisabled={state === SilenceStates.Expired}
-            onChange={onCheckboxChange}
-          />
-        </td>
-      )}
-      <td className={tableSilenceClasses[1]}>
-        <div className="co-resource-item">
-          <MonitoringResourceIcon resource={SilenceResource} />
-          <Link
-            className="co-resource-item__resource-name"
-            data-test-id="silence-resource-link"
-            title={id}
-            to={`${SilenceResource.plural}/${id}`}
-          >
-            {name}
-          </Link>
-        </div>
-        <div className="monitoring-label-list">
-          <SilenceMatchersList silence={obj} />
-        </div>
-      </td>
-      <td className={tableSilenceClasses[2]}>
-        <SeverityCounts alerts={firingAlerts} />
-      </td>
-      <td className={classNames(tableSilenceClasses[3], 'co-break-word')}>
-        <SilenceState silence={obj} />
-        {state === SilenceStates.Pending && (
-          <StateTimestamp text={t('Starts')} timestamp={startsAt} />
-        )}
-        {state === SilenceStates.Active && <StateTimestamp text={t('Ends')} timestamp={endsAt} />}
-        {state === SilenceStates.Expired && (
-          <StateTimestamp text={t('Expired')} timestamp={endsAt} />
-        )}
-      </td>
-      <td className={tableSilenceClasses[4]}>{createdBy || '-'}</td>
-      <td className={tableSilenceClasses[5]}>
-        <SilenceDropdownKebab silence={obj} />
-      </td>
-    </>
-  );
-};
-
 const SilenceTableRowWithCheckbox: React.FC<RowProps<Silence>> = ({ obj }) => (
   <SilenceTableRow showCheckbox={true} obj={obj} />
 );
-
-const alertMessageResources: {
-  [labelName: string]: { kind: string; namespaced?: boolean };
-} = {
-  container: ContainerModel,
-  daemonset: DaemonSetModel,
-  deployment: DeploymentModel,
-  job: JobModel,
-  namespace: NamespaceModel,
-  node: NodeModel,
-  pod: PodModel,
-  statefulset: StatefulSetModel,
-};
-
-const matchCount = (haystack: string, regExpString: string) =>
-  _.size(haystack.match(new RegExp(regExpString, 'g')));
-
-const AlertMessage: React.FC<AlertMessageProps> = ({ alertText, labels, template }) => {
-  if (_.isEmpty(alertText)) {
-    return null;
-  }
-
-  let messageParts: React.ReactNode[] = [alertText];
-
-  // Go through each recognized resource type and replace any resource names that exist in alertText
-  // with a link to the resource's details page
-  _.each(alertMessageResources, (model, label) => {
-    const labelValue = labels[label];
-
-    if (labelValue && !(model.namespaced && _.isEmpty(labels.namespace))) {
-      const tagCount = matchCount(template, `\\{\\{ *\\$labels\\.${label} *\\}\\}`);
-      const resourceNameCount = matchCount(alertText, _.escapeRegExp(labelValue));
-
-      // Don't do the replacement unless the counts match. This avoids overwriting the wrong string
-      // if labelValue happens to appear elsewhere in alertText
-      if (tagCount > 0 && tagCount === resourceNameCount) {
-        const link = (
-          <ResourceLink
-            className="co-resource-item--monitoring-alert"
-            inline
-            key={model.kind}
-            kind={model.kind}
-            name={labelValue}
-            namespace={model.namespaced ? labels.namespace : undefined}
-          />
-        );
-        messageParts = _.flatMap(messageParts, (part) => {
-          if (_.isString(part) && part.indexOf(labelValue) !== -1) {
-            // `part` contains at least one instance of the resource name, so replace each instance
-            // with the link to the resource. Since the link is a component, we can't simply do a
-            // string substitution. Instead, create an array that contains each of the string parts
-            // and the resource links in the correct order.
-            const splitParts = part.split(labelValue);
-            return _.flatMap(splitParts, (p) => [p, link]).slice(0, -1);
-          }
-          return [part];
-        });
-      }
-    }
-  });
-
-  return (
-    <div className="co-alert-manager">
-      <p>
-        <LinkifyExternal>{messageParts}</LinkifyExternal>
-      </p>
-    </div>
-  );
-};
-
-const HeaderAlertMessage: React.FC<{ alert: Alert; rule: Rule }> = ({ alert, rule }) => {
-  const annotation = alert.annotations.description ? 'description' : 'message';
-  return (
-    <AlertMessage
-      alertText={alert.annotations[annotation]}
-      labels={alert.labels}
-      template={rule.annotations[annotation]}
-    />
-  );
-};
-
-const getSourceKey = (source, t: TFunction) => {
-  switch (source) {
-    case 'Platform':
-      return t('Platform');
-    case 'User':
-      return t('User');
-    default:
-      return source;
-  }
-};
-
-const SilencedByList: React.FC<{ silences: Silence[] }> = ({ silences }) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  const columns = React.useMemo<TableColumn<Silence>[]>(
-    () => [
-      {
-        id: 'name',
-        props: { className: tableSilenceClasses[1] },
-        title: t('Name'),
-      },
-      {
-        id: 'firingAlerts',
-        props: { className: tableSilenceClasses[2] },
-        title: t('Firing alerts'),
-      },
-      {
-        id: 'state',
-        props: { className: tableSilenceClasses[3] },
-        title: t('State'),
-      },
-      {
-        id: 'createdBy',
-        props: { className: tableSilenceClasses[4] },
-        title: t('Creator'),
-      },
-      {
-        id: 'actions',
-        props: { className: tableSilenceClasses[5] },
-        title: '',
-      },
-    ],
-    [t],
-  );
-
-  return (
-    <VirtualizedTable<Silence>
-      aria-label={t('Silenced by')}
-      columns={columns}
-      data={silences}
-      loaded={true}
-      loadError={undefined}
-      Row={SilenceTableRow}
-      unfilteredData={silences}
-    />
-  );
-};
-
-type AlertsDetailsPageProps = RouteComponentProps<{ ns?: string; ruleID: string }>;
-
-const AlertsDetailsPage_: React.FC<AlertsDetailsPageProps> = ({ history, match }) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  const { alertsKey } = usePerspective();
-  const namespace = match.params?.ns;
-  const hideGraphs = useSelector(({ observe }: RootState) => !!observe.get('hideGraphs'));
-
-  const alerts: Alerts = useSelector(({ observe }: RootState) => observe.get(alertsKey));
-
-  const silencesLoaded = ({ observe }) => observe.get('silences')?.loaded;
-
-  const ruleAlerts = _.filter(alerts?.data, (a) => a.rule.id === match?.params?.ruleID);
-  const rule = ruleAlerts?.[0]?.rule;
-
-  // Search for an alert that matches all of the labels in the URL parameters. We expect there to be
-  // only one such alert that matches, so don't display any alert if multiple matches were found.
-  const queryParams = getAllQueryArguments();
-  const foundAlerts = _.filter(ruleAlerts, (a) => _.isMatch(a.labels, queryParams));
-  const alert = foundAlerts.length === 1 ? foundAlerts[0] : undefined;
-
-  const state = alertState(alert);
-
-  const labelsMemoKey = JSON.stringify(alert?.labels);
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  const labels: PrometheusLabels = React.useMemo(() => alert?.labels, [labelsMemoKey]);
-
-  // eslint-disable-next-line camelcase
-  const runbookURL = alert?.annotations?.runbook_url;
-
-  const sourceId = rule?.sourceId;
-
-  // Load alert metrics chart from plugin
-  const [alertsChartExtensions] =
-    useResolvedExtensions<AlertingRuleChartExtension>(isAlertingRuleChart);
-  const alertsChart = alertsChartExtensions
-    .filter((extension) => extension.properties.sourceId === sourceId)
-    .map((extension) => extension.properties.chart);
-
-  const AlertsChart = alertsChart?.[0];
-
-  return (
-    <>
-      <StatusBox
-        data={alert}
-        label={AlertResource.label}
-        loaded={alerts?.loaded}
-        loadError={alerts?.loadError}
-      >
-        <div className="pf-c-page__main-breadcrumb">
-          <Breadcrumb className="monitoring-breadcrumbs">
-            <BreadcrumbItem>
-              <Link
-                className="pf-c-breadcrumb__link"
-                to={namespace ? `/dev-monitoring/ns/${namespace}/alerts` : '/monitoring/alerts'}
-              >
-                {t('Alerts')}
-              </Link>
-            </BreadcrumbItem>
-            <BreadcrumbItem isActive>{t('Alert details')}</BreadcrumbItem>
-          </Breadcrumb>
-        </div>
-        <div className="co-m-nav-title co-m-nav-title--detail co-m-nav-title--breadcrumbs">
-          <h1 className="co-m-pane__heading">
-            <div data-test="resource-title" className="co-resource-item">
-              <MonitoringResourceIcon className="co-m-resource-icon--lg" resource={AlertResource} />
-              {labels?.alertname}
-              <SeverityBadge severity={labels?.severity} />
-            </div>
-            {state !== AlertStates.Silenced && (
-              <div data-test-id="details-actions">
-                <Button
-                  className="co-action-buttons__btn"
-                  onClick={() => history.push(silenceAlertURL(alert))}
-                  variant="primary"
-                >
-                  {t('Silence alert')}
-                </Button>
-              </div>
-            )}
-          </h1>
-          <HeaderAlertMessage alert={alert} rule={rule} />
-        </div>
-        <div className="co-m-pane__body">
-          <Toolbar className="monitoring-alert-detail-toolbar">
-            <ToolbarContent>
-              <ToolbarItem variant="label">
-                <SectionHeading text={t('Alert details')} />
-              </ToolbarItem>
-              <ToolbarGroup alignment={{ default: 'alignRight' }}>
-                <ActionServiceProvider context={{ 'alert-detail-toolbar-actions': { alert } }}>
-                  {({ actions, loaded }) =>
-                    loaded
-                      ? actions.map((action) => {
-                          if (isActionWithHref(action)) {
-                            return (
-                              <ToolbarItem
-                                key={action.id}
-                                spacer={{ default: 'spacerNone' }}
-                                className="pf-u-px-md"
-                              >
-                                <Link to={action.cta.href}>{action.label}</Link>
-                              </ToolbarItem>
-                            );
-                          } else if (isActionWithCallback(action)) {
-                            return (
-                              <ToolbarItem key={action.id} spacer={{ default: 'spacerNone' }}>
-                                <Button variant="link" onClick={action.cta}>
-                                  {action.label}
-                                </Button>
-                              </ToolbarItem>
-                            );
-                          }
-
-                          return null;
-                        })
-                      : null
-                  }
-                </ActionServiceProvider>
-                <ToolbarItem>
-                  <ToggleGraph />
-                </ToolbarItem>
-              </ToolbarGroup>
-            </ToolbarContent>
-          </Toolbar>
-
-          <div className="co-m-pane__body-group">
-            <div className="row">
-              <div className="col-sm-12">
-                {!sourceId || sourceId === 'prometheus' ? (
-                  <Graph
-                    filterLabels={labels}
-                    namespace={namespace}
-                    query={rule?.query}
-                    ruleDuration={rule?.duration}
-                  />
-                ) : AlertsChart && !hideGraphs ? (
-                  <AlertsChart rule={rule} />
-                ) : null}
-              </div>
-            </div>
-            <div className="row">
-              <div className="col-sm-6">
-                <dl className="co-m-pane__details">
-                  <dt>{t('Name')}</dt>
-                  <dd>{labels?.alertname}</dd>
-                  <dt>
-                    <PopoverField bodyContent={<SeverityHelp />} label={t('Severity')} />
-                  </dt>
-                  <dd>
-                    <Severity severity={labels?.severity} />
-                  </dd>
-                  {alert?.annotations?.description && (
-                    <>
-                      <dt>{t('Description')}</dt>
-                      <dd>
-                        <AlertMessage
-                          alertText={alert.annotations.description}
-                          labels={labels}
-                          template={rule?.annotations?.description}
-                        />
-                      </dd>
-                    </>
-                  )}
-                  {alert?.annotations?.summary && (
-                    <>
-                      <dt>{t('Summary')}</dt>
-                      <dd>{alert.annotations.summary}</dd>
-                    </>
-                  )}
-                  {alert?.annotations?.message && (
-                    <>
-                      <dt>{t('Message')}</dt>
-                      <dd>
-                        <AlertMessage
-                          alertText={alert.annotations.message}
-                          labels={labels}
-                          template={rule?.annotations?.message}
-                        />
-                      </dd>
-                    </>
-                  )}
-                  {runbookURL && (
-                    <>
-                      <dt>{t('Runbook')}</dt>
-                      <dd>
-                        <ExternalLink href={runbookURL} text={runbookURL} />
-                      </dd>
-                    </>
-                  )}
-                </dl>
-              </div>
-              <div className="col-sm-6">
-                <dl className="co-m-pane__details">
-                  <dt>
-                    <PopoverField bodyContent={<SourceHelp />} label={t('Source')} />
-                  </dt>
-                  <dd>{alert && getSourceKey(_.startCase(alertSource(alert)), t)}</dd>
-                  <dt>
-                    <PopoverField bodyContent={<AlertStateHelp />} label={t('State')} />
-                  </dt>
-                  <dd>
-                    <AlertState state={state} />
-                    <AlertStateDescription alert={alert} />
-                  </dd>
-                </dl>
-              </div>
-            </div>
-          </div>
-          <div className="co-m-pane__body-group">
-            <div className="row">
-              <div className="col-xs-12">
-                <dl className="co-m-pane__details" data-test="label-list">
-                  <dt>{t('Labels')}</dt>
-                  <dd>
-                    <Labels kind="alert" labels={labels} />
-                  </dd>
-                </dl>
-              </div>
-            </div>
-          </div>
-          <div className="co-m-pane__body-group">
-            <div className="row">
-              <div className="col-xs-12">
-                <dl className="co-m-pane__details">
-                  <dt>{t('Alerting rule')}</dt>
-                  <dd>
-                    <div className="co-resource-item">
-                      <MonitoringResourceIcon resource={RuleResource} />
-                      <Link
-                        to={namespace ? devRuleURL(rule, namespace) : ruleURL(rule)}
-                        data-test="alert-rules-detail-resource-link"
-                        className="co-resource-item__resource-name"
-                      >
-                        {_.get(rule, 'name')}
-                      </Link>
-                    </div>
-                  </dd>
-                </dl>
-              </div>
-            </div>
-          </div>
-        </div>
-        {silencesLoaded && !_.isEmpty(alert?.silencedBy) && (
-          <div className="co-m-pane__body">
-            <div className="co-m-pane__body-group">
-              <SectionHeading text={t('Silenced by')} />
-              <div className="row">
-                <div className="col-xs-12">
-                  <SilencedByList silences={alert?.silencedBy} />
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-      </StatusBox>
-    </>
-  );
-};
-const AlertsDetailsPage = withFallback(withRouter(AlertsDetailsPage_));
 
 // Renders Prometheus template text and highlights any {{ ... }} tags that it contains
 const PrometheusTemplate = ({ text }) => (
@@ -1159,143 +427,6 @@ const AlertRulesDetailsPage_: React.FC<AlertRulesDetailsPageProps> = ({ match })
   );
 };
 export const AlertRulesDetailsPage = withFallback(AlertRulesDetailsPage_);
-
-type ExpireSilenceModalProps = {
-  isOpen: boolean;
-  setClosed: () => void;
-  silenceID: string;
-};
-
-const ExpireSilenceModal: React.FC<ExpireSilenceModalProps> = ({
-  isOpen,
-  setClosed,
-  silenceID,
-}) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-  const { perspective } = usePerspective();
-
-  const dispatch = useDispatch();
-
-  const [isInProgress, , setInProgress, setNotInProgress] = useBoolean(false);
-  const [errorMessage, setErrorMessage] = React.useState();
-
-  const expireSilence = () => {
-    setInProgress();
-    consoleFetchJSON
-      .delete(`${window.SERVER_FLAGS.alertManagerBaseURL}/api/v2/silence/${silenceID}`)
-      .then(() => {
-        refreshSilences(dispatch, perspective);
-        setClosed();
-      })
-      .catch((err) => {
-        setErrorMessage(_.get(err, 'json.error') || err.message || 'Error expiring silence');
-        setNotInProgress();
-      })
-      .then(setNotInProgress);
-  };
-
-  return (
-    <Modal
-      isOpen={isOpen}
-      position="top"
-      showClose={false}
-      title={t('Expire silence')}
-      variant={ModalVariant.small}
-    >
-      <Flex direction={{ default: 'column' }}>
-        <FlexItem>{t('Are you sure you want to expire this silence?')}</FlexItem>
-        <Flex direction={{ default: 'column' }}>
-          <FlexItem>
-            {errorMessage && (
-              <PFAlert
-                className="co-alert co-alert--scrollable"
-                isInline
-                title={t('An error occurred')}
-                variant="danger"
-              >
-                <div className="co-pre-line">{errorMessage}</div>
-              </PFAlert>
-            )}
-          </FlexItem>
-          <Flex>
-            <FlexItem>{isInProgress && <LoadingInline />}</FlexItem>
-            <FlexItem align={{ default: 'alignRight' }}>
-              <Button variant="secondary" onClick={setClosed}>
-                {t('Cancel')}
-              </Button>
-            </FlexItem>
-            <FlexItem>
-              <Button variant="primary" onClick={expireSilence}>
-                {t('Expire silence')}
-              </Button>
-            </FlexItem>
-          </Flex>
-        </Flex>
-      </Flex>
-    </Modal>
-  );
-};
-
-type SilenceDropdownProps = RouteComponentProps & {
-  className?: string;
-  isPlain?: boolean;
-  silence: Silence;
-  Toggle: React.FC<{ onToggle: OnToggle }>;
-};
-
-const SilenceDropdown_: React.FC<SilenceDropdownProps> = ({
-  className,
-  history,
-  isPlain,
-  silence,
-  Toggle,
-}) => {
-  const { t } = useTranslation('plugin__monitoring-plugin');
-
-  const [isOpen, setIsOpen, , setClosed] = useBoolean(false);
-  const [isModalOpen, , setModalOpen, setModalClosed] = useBoolean(false);
-
-  const editSilence = () => {
-    history.push(`${SilenceResource.plural}/${silence.id}/edit`);
-  };
-
-  const dropdownItems =
-    silenceState(silence) === SilenceStates.Expired
-      ? [
-          <DropdownItem key="edit-silence" component="button" onClick={editSilence}>
-            {t('Recreate silence')}
-          </DropdownItem>,
-        ]
-      : [
-          <DropdownItem key="edit-silence" component="button" onClick={editSilence}>
-            {t('Edit silence')}
-          </DropdownItem>,
-          <DropdownItem key="cancel-silence" component="button" onClick={setModalOpen}>
-            {t('Expire silence')}
-          </DropdownItem>,
-        ];
-
-  return (
-    <>
-      <Dropdown
-        className={className}
-        data-test="silence-actions"
-        dropdownItems={dropdownItems}
-        isOpen={isOpen}
-        isPlain={isPlain}
-        onSelect={setClosed}
-        position={DropdownPosition.right}
-        toggle={<Toggle onToggle={setIsOpen} />}
-      />
-      <ExpireSilenceModal isOpen={isModalOpen} setClosed={setModalClosed} silenceID={silence.id} />
-    </>
-  );
-};
-const SilenceDropdown = withRouter(SilenceDropdown_);
-
-const SilenceDropdownKebab: React.FC<{ silence: Silence }> = ({ silence }) => (
-  <SilenceDropdown isPlain silence={silence} Toggle={KebabToggle} />
-);
 
 const ActionsToggle: React.FC<{ onToggle: OnToggle }> = ({ onToggle, ...props }) => (
   <DropdownToggle data-test="silence-actions-toggle" onToggle={onToggle} {...props}>
@@ -2060,22 +1191,5 @@ const MonitoringUI = () => (
     <Route component={PollerPages} />
   </Switch>
 );
-
-type AlertMessageProps = {
-  alertText: string;
-  labels: PrometheusLabels;
-  template: string;
-};
-
-type GraphProps = {
-  filterLabels?: PrometheusLabels;
-  formatSeriesTitle?: FormatSeriesTitle;
-  namespace?: string;
-  query: string;
-  ruleDuration: number;
-  showLegend?: boolean;
-};
-
-type OnToggle = (value: boolean, e: MouseEvent) => void;
 
 export default MonitoringUI;
